Add CLI entry point for Streamlit app
