+++ README.md
@@ -1,7 +1,22 @@
 # WebLEDMatrix
 
-```shell
+A small Streamlit application for controlling LED matrix controllers.
+
+## Installation
+
+The package can be installed from source using Poetry or pip:
+
+```bash
 poetry install
 poetry build
 pip install dist/
+```
+
+## Usage
+
+After installation a command line entry point named `web-led-matrix` is available.
+It launches the Streamlit interface for selecting and identifying matrices:
+
+```bash
+web-led-matrix
 ```
