+++ pyproject.toml
@@ -12,6 +12,9 @@ dependencies = [
     "streamlit (>=1.46.0,<2.0.0)"
 ]
 
+[project.scripts]
+web-led-matrix = "web_led_matrix.cli:main"
+
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
